requirements.txt
streamlit==1.32.0
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit-autorefresh==1.0.1
sounddevice==0.4.6
numpy==1.26.3
spotipy==2.23.0
yfinance==0.2.36
plotly==5.18.0
streamlit-extras==0.2.1
pillow==10.2.0
